requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncer==0.0.1
brotlipy==0.7.0
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.5
cycler==0.11.0
fastapi==0.80.0
filelock==3.8.0
filetype==1.1.0
flatbuffers==22.10.26
fonttools==4.38.0
freetype-py==2.3.0
frozenlist==1.3.1
gdown==4.5.1
h11==0.14.0
humanfriendly==10.0
ImageHash==4.2.1
imageio==2.22.2
iopath==0.1.9
joblib==1.2.0
kaolin==0.12.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.1
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.7
numba==0.56.4
numpy==1.22.4
onnxruntime-gpu==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
Pillow==9.2.0
ply==3.11
protobuf==4.21.9
PyQt5-sip==12.11.0
pyrender==0.1.45
python-multipart==0.0.5
pytorch3d==0.7.1
PyWavelets==1.4.1
rembg==2.0.25
Rtree==1.0.1
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.7.2
smplx==0.1.28
starlette==0.19.1
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
trimesh==3.15.7
yarl==1.8.1

